use percentages rather than vector normalization
Using vector normalization (i.e. dividing by the magnitude of the vector) doesn't actually make very much sense for the decision tree. If we think about things like checking for semicolons, what we really want to know is if more than X% of the tokens are semicolons; we don't care about the semicolon component of a normalized vector.